move fullstack types to dedicated source
