+++ pkg/fullstack/gcp/fullstack.go
@@ -7,115 +7,8 @@ import (
 	"strings"
 
 	apigateway "github.com/pulumi/pulumi-gcp/sdk/v8/go/gcp/apigateway"
-	"github.com/pulumi/pulumi-gcp/sdk/v8/go/gcp/serviceaccount"
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
-
-	cloudrunv2 "github.com/pulumi/pulumi-gcp/sdk/v8/go/gcp/cloudrunv2"
-	"github.com/pulumi/pulumi-gcp/sdk/v8/go/gcp/compute"
-	"github.com/pulumi/pulumi-gcp/sdk/v8/go/gcp/dns"
 )
-
-// FullStack represents a complete fullstack application infrastructure on Google Cloud Platform.
-type FullStack struct {
-	pulumi.ResourceState
-
-	Project       string
-	Region        string
-	BackendName   string
-	BackendImage  pulumi.StringOutput
-	FrontendName  string
-	FrontendImage pulumi.StringOutput
-	Labels        map[string]string
-
-	name string
-
-	backendService  *cloudrunv2.Service
-	backendAccount  *serviceaccount.Account
-	frontendService *cloudrunv2.Service
-	frontendAccount *serviceaccount.Account
-	apiGateway      *apigateway.Gateway
-	apiConfig       *apigateway.ApiConfig
-
-	// IAM members for API Gateway invoker permissions
-	gatewayServiceAccount    *serviceaccount.Account
-	backendGatewayIamMember  *cloudrunv2.ServiceIamMember
-	frontendGatewayIamMember *cloudrunv2.ServiceIamMember
-
-	// Network infrastructure
-	certificate          *compute.ManagedSslCertificate
-	neg                  *compute.RegionNetworkEndpointGroup
-	globalForwardingRule *compute.GlobalForwardingRule
-	dnsRecord            *dns.RecordSet
-}
-
-// FullStackArgs contains configuration arguments for creating a FullStack instance.
-type FullStackArgs struct {
-	Project       string
-	Region        string
-	BackendName   string
-	BackendImage  pulumi.StringInput
-	FrontendName  string
-	FrontendImage pulumi.StringInput
-	// Optional additional config
-	Backend  *BackendArgs
-	Frontend *FrontendArgs
-	Network  *NetworkArgs
-	Labels   map[string]string
-}
-
-// BackendArgs contains configuration for the backend service.
-type BackendArgs struct {
-	*InstanceArgs
-}
-
-// FrontendArgs contains configuration for the frontend service.
-type FrontendArgs struct {
-	*InstanceArgs
-	EnableUnauthenticated bool
-}
-
-// Probe contains configuration for health check probes
-type Probe struct {
-	Path                string
-	InitialDelaySeconds int
-	PeriodSeconds       int
-	TimeoutSeconds      int
-	FailureThreshold    int
-}
-
-// InstanceArgs contains configuration for Cloud Run service instances.
-type InstanceArgs struct {
-	ResourceLimits       pulumi.StringMap
-	SecretConfigFileName string
-	SecretConfigFilePath string
-	EnvVars              map[string]string
-	MaxInstanceCount     int
-	DeletionProtection   bool
-	ContainerPort        int
-	StartupProbe         *Probe
-	LivenessProbe        *Probe
-}
-
-// NetworkArgs contains configuration for network infrastructure including load balancers and API Gateway.
-type NetworkArgs struct {
-	// Domain name for the internet-facing certificate. Required.
-	// E.g.: "myapp.path2prod.dev"
-	DomainURL string
-	// GCP network where to host the load balancer instances. Defaults to "default".
-	ProxyNetworkName string
-	// Whether to apply best-practice Cloud Armor policies to the load balancer. Defaults to false.
-	EnableCloudArmor bool
-	// Whether to enable Identity Aware Proxy for authentication. Defaults to false.
-	EnableIAP bool
-	// Support email for IAP's OAuth consent screen. Required if EnableIAP=true.
-	IAPSupportEmail string
-	// Whether to restrict access to the given list of client IPs. Valid only when EnableCloudArmor=true.
-	ClientIPAllowlist []string
-	// Whether to disable public internet access. Useful during development. Defaults to false.
-	EnablePrivateTrafficOnly bool
-	// API Gateway configuration. If provided, traffic will be routed through API Gateway.
-	APIGateway *APIGatewayArgs
-}
 
 // NewFullStack creates a new FullStack instance with the provided configuration.
 func NewFullStack(ctx *pulumi.Context, name string, args *FullStackArgs, opts ...pulumi.ResourceOption) (*FullStack, error) {
@@ -280,108 +173,4 @@ func (f *FullStack) newResourceName(serviceName, resourceType string, maxLength 
 	}
 
 	return resourceName
-}
-
-// applyDefaultGatewayArgs applies default API Gateway configuration to the provided args.
-// If the provided args is nil, it returns a new instance with default config.
-// If the provided args has a nil Config, it applies the default config.
-func applyDefaultGatewayArgs(args *APIGatewayArgs, backendServiceURL, frontendServiceURL pulumi.StringOutput) *APIGatewayArgs {
-	var gatewayArgs *APIGatewayArgs
-	if args == nil {
-		gatewayArgs = &APIGatewayArgs{}
-	} else {
-		gatewayArgs = args
-	}
-
-	if gatewayArgs.Config == nil {
-		gatewayArgs.Config = &APIConfigArgs{}
-	}
-
-	// Initialize Backend and Frontend if they are nil
-	if gatewayArgs.Config.Backend == nil {
-		gatewayArgs.Config.Backend = &Upstream{}
-	}
-	if gatewayArgs.Config.Frontend == nil {
-		gatewayArgs.Config.Frontend = &Upstream{}
-	}
-
-	// Ignore any value given and set always to the services URLs
-	gatewayArgs.Config.Backend.ServiceURL = backendServiceURL
-	gatewayArgs.Config.Frontend.ServiceURL = frontendServiceURL
-
-	if gatewayArgs.Name == "" {
-		gatewayArgs.Name = "gateway"
-	}
-
-	return gatewayArgs
-}
-
-// GetBackendService returns the backend Cloud Run service.
-func (f *FullStack) GetBackendService() *cloudrunv2.Service {
-	return f.backendService
-}
-
-// GetFrontendService returns the frontend Cloud Run service.
-func (f *FullStack) GetFrontendService() *cloudrunv2.Service {
-	return f.frontendService
-}
-
-// GetAPIGateway returns the API Gateway instance.
-func (f *FullStack) GetAPIGateway() *apigateway.Gateway {
-	return f.apiGateway
-}
-
-// GetAPIConfig returns the API Gateway configuration.
-func (f *FullStack) GetAPIConfig() *apigateway.ApiConfig {
-	return f.apiConfig
-}
-
-// GetBackendGatewayIamMember returns the backend service IAM member for API Gateway invoker permissions.
-func (f *FullStack) GetBackendGatewayIamMember() *cloudrunv2.ServiceIamMember {
-	return f.backendGatewayIamMember
-}
-
-// GetFrontendGatewayIamMember returns the frontend service IAM member for API Gateway invoker permissions.
-func (f *FullStack) GetFrontendGatewayIamMember() *cloudrunv2.ServiceIamMember {
-	return f.frontendGatewayIamMember
-}
-
-// GetCertificate returns the managed SSL certificate for the domain.
-func (f *FullStack) GetCertificate() *compute.ManagedSslCertificate {
-	return f.certificate
-}
-
-// GetNEG returns the region network endpoint group for the load balancer.
-func (f *FullStack) GetNEG() *compute.RegionNetworkEndpointGroup {
-	return f.neg
-}
-
-// GetGlobalForwardingRule returns the global forwarding rule for the load balancer.
-func (f *FullStack) GetGlobalForwardingRule() *compute.GlobalForwardingRule {
-	return f.globalForwardingRule
-}
-
-// LookupDNSZone finds the appropriate DNS managed zone for the given domain in the current project
-func (f *FullStack) LookupDNSZone(ctx *pulumi.Context, domainURL string) (string, error) {
-	return f.lookupDNSZone(ctx, domainURL, f.Project)
-}
-
-// GetDNSRecord returns the DNS record created for the load balancer
-func (f *FullStack) GetDNSRecord() *dns.RecordSet {
-	return f.dnsRecord
-}
-
-// GetBackendAccount returns the backend service account.
-func (f *FullStack) GetBackendAccount() *serviceaccount.Account {
-	return f.backendAccount
-}
-
-// GetFrontendAccount returns the frontend service account.
-func (f *FullStack) GetFrontendAccount() *serviceaccount.Account {
-	return f.frontendAccount
-}
-
-// GetGatewayServiceAccount returns the API Gateway service account.
-func (f *FullStack) GetGatewayServiceAccount() *serviceaccount.Account {
-	return f.gatewayServiceAccount
 }
